clean up, renamed search result list
